+++ source/Nuke.Common/CI/Bitrise/BitriseOutputSink.cs
@@ -20,11 +20,5 @@ namespace Nuke.Common.CI.Bitrise
         protected override string WarningCode => "33;1";
         protected override string ErrorCode => "31;1";
         protected override string SuccessCode => "32;1";
-
-        internal override IDisposable WriteBlock(string text)
-        {
-            return DelegateDisposable.CreateBracket(
-                () => WriteNormal(FigletTransform.GetText(text, "ansi-shadow")));
-        }
     }
 }