+++ source/Nuke.Common/Execution/ExecutableTarget.cs
@@ -8,6 +8,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
+using Nuke.Common.Tooling;
 using Nuke.Common.Utilities.Collections;
 // ReSharper disable MissingBaseTypeHighlighting
 
@@ -38,6 +39,10 @@ namespace Nuke.Common.Execution
 
         public IReadOnlyCollection<ExecutableTarget> AllDependencies
             => ExecutionDependencies.Concat(OrderDependencies).Concat(TriggerDependencies).ToList();
+
+        public int? PartitionSize { get; set; }
+        public LookupTable<ExecutableTarget, string> ArtifactDependencies { get; internal set; } = new LookupTable<ExecutableTarget, string>();
+        public ICollection<string> ArtifactProducts { get; internal set; } = new List<string>();
 
         public TimeSpan Duration => Stopwatch.Elapsed;
         public bool IsDefault { get; set; }