+++ source/Nuke.Common/INukeBuild.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
+using Nuke.Common.CI;
 using Nuke.Common.Execution;
 using Nuke.Common.IO;
 using Nuke.Common.Tooling;
@@ -45,5 +46,6 @@ namespace Nuke.Common
         bool IsServerBuild {get;}
         LogLevel LogLevel {get;}
         bool Continue { get; }
+        Partition Partition { get; }
     }
 }