+++ source/Nuke.Common/Utilities/Collections/Dictionary.AddDictionary.cs
@@ -8,9 +8,20 @@ namespace Nuke.Common.Utilities.Collections
 {
     public static partial class DictionaryExtensions
     {
-        public static Dictionary<TKey, TValue> AddDictionary<TKey, TValue>(
-            this Dictionary<TKey, TValue> dictionary,
+        public static TDictionary AddDictionary<TDictionary, TKey, TValue>(
+            this TDictionary dictionary,
+            IDictionary<TKey, TValue> otherDictionary)
+            where TDictionary : IDictionary<TKey, TValue>
+        {
+            foreach (var (key, value) in otherDictionary)
+                dictionary.AddPair(key, value);
+            return dictionary;
+        }
+
+        public static TDictionary AddDictionary<TDictionary, TKey, TValue>(
+            this TDictionary dictionary,
             IReadOnlyDictionary<TKey, TValue> otherDictionary)
+            where TDictionary : IDictionary<TKey, TValue>
         {
             foreach (var (key, value) in otherDictionary)
                 dictionary.AddPair(key, value);