+++ source/Nuke.SourceGenerators.Tests/StronglyTypedSolutionGeneratorTest.cs
@@ -3,9 +3,7 @@
 // https://github.com/nuke-build/nuke/blob/master/LICENSE
 
 using System;
-using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.CodeAnalysis;
@@ -33,12 +31,9 @@ partial class Build : NukeBuild
 }");
 
             var generator = new StronglyTypedSolutionGenerator();
-            var driver = (GeneratorDriver) CSharpGeneratorDriver.Create(generator);
-            var result = driver.RunGeneratorsAndUpdateCompilation(inputCompilation, out _, out _).GetRunResult();
-
-            if (!result.Diagnostics.IsEmpty)
-                throw new Exception(string.Join(Environment.NewLine, result.Diagnostics.Select(x => x.GetMessage())));
-            return Verifier.Verify(result.GeneratedTrees.Single().GetRoot().ToFullString());
+            var driver = CSharpGeneratorDriver.Create(generator);
+            var result = driver.RunGenerators(inputCompilation);
+            return Verifier.Verify(result);
         }
 
         [Fact]
@@ -55,8 +50,8 @@ partial class Build : NukeBuild
 }");
 
             var generator = new StronglyTypedSolutionGenerator();
-            var driver = (GeneratorDriver) CSharpGeneratorDriver.Create(generator);
-            var result = driver.RunGeneratorsAndUpdateCompilation(inputCompilation, out _, out _).GetRunResult();
+            var driver = CSharpGeneratorDriver.Create(generator);
+            var result = driver.RunGenerators(inputCompilation).GetRunResult();
 
             if (!result.Diagnostics.IsEmpty)
                 throw new Exception(string.Join(Environment.NewLine, result.Diagnostics.Select(x => x.GetMessage())));
@@ -77,8 +72,8 @@ partial class Build : NukeBuild
 }");
 
             var generator = new StronglyTypedSolutionGenerator();
-            var driver = (GeneratorDriver) CSharpGeneratorDriver.Create(generator);
-            var result = driver.RunGeneratorsAndUpdateCompilation(inputCompilation, out _, out _).GetRunResult();
+            var driver = CSharpGeneratorDriver.Create(generator);
+            var result = driver.RunGenerators(inputCompilation).GetRunResult();
 
             if (!result.Diagnostics.IsEmpty)
                 throw new Exception(string.Join(Environment.NewLine, result.Diagnostics.Select(x => x.GetMessage())));
@@ -89,13 +84,8 @@ partial class Build : NukeBuild
         {
             return CSharpCompilation.Create("compilation",
                 new[] { CSharpSyntaxTree.ParseText(source) },
-                new[]
-                {
-                    MetadataReference.CreateFromFile(typeof(object).GetTypeInfo().Assembly.Location),
-                    MetadataReference.CreateFromFile(Path.Combine(Path.GetDirectoryName(typeof(object).Assembly.Location)!, "netstandard.dll")),
-                    MetadataReference.CreateFromFile(Path.Combine(Path.GetDirectoryName(typeof(object).Assembly.Location)!, "System.Runtime.dll")),
-                    MetadataReference.CreateFromFile(typeof(SolutionAttribute).Assembly.Location)
-                },
+                new[] { MetadataReference.CreateFromFile(typeof(SolutionAttribute).Assembly.Location) }
+                    .Concat(Basic.Reference.Assemblies.NetStandard20.All),
                 new CSharpCompilationOptions(OutputKind.ConsoleApplication));
         }
     }